+++ UNIWA EEE.2.5.2 Movie Theater Project.API/Services/MovieTheater/MovieTheaterService.cs
@@ -4,16 +4,21 @@ using FluentValidation.Results;
 using MovieTheaterProject.API.Domain;
 using MovieTheaterProject.API.Mapping;
 using MovieTheaterProject.API.Repositories.MovieTheater;
+using MovieTheaterProject.API.Repositories.MovieViewing;
 
 namespace MovieTheaterProject.API.Services.MovieTheater;
 
 public sealed class MovieTheaterService : IMovieTheaterService
 {
     private readonly IMovieTheaterRepository _movieTheaterRepository;
+    private readonly IMovieViewingRepository _movieViewingRepository;
 
-    public MovieTheaterService(IMovieTheaterRepository movieTheaterRepository)
+    public MovieTheaterService(
+        IMovieTheaterRepository movieTheaterRepository,
+        IMovieViewingRepository movieViewingRepository)
     {
         _movieTheaterRepository = movieTheaterRepository;
+        _movieViewingRepository = movieViewingRepository;
     }
 
     public async Task<bool> CreateAsync(Domain.MovieTheater movieTheater)
@@ -24,7 +29,7 @@ public sealed class MovieTheaterService : IMovieTheaterService
             var message = $"A movie theater with the id {movieTheater.Id} already exists";
             throw new ValidationException(message, new[]
             {
-                new ValidationFailure(nameof(Movie), message)
+                new ValidationFailure(nameof(MovieTheater), message)
             });
         }
 
@@ -52,6 +57,15 @@ public sealed class MovieTheaterService : IMovieTheaterService
 
     public async Task<bool> DeleteAsync(Guid id)
     {
+        if ((await _movieViewingRepository.GetAsyncByMovieTheaterId(id)).Count() != 0)
+        {
+            var message = $"At least a movie viewing is scheduled for the movie theater with the id {id}";
+            throw new ValidationException(message, new[]
+            {
+                new ValidationFailure(nameof(MovieTheater), message)
+            });
+        }
+
         return await _movieTheaterRepository.DeleteAsync(id);
     }
 }
